feat(web): convert /directory and /report/[domain] to hono/jsx
- directory: Row component, ld+json via raw(), dofollow backlinks + entities.
- report/[domain]: print-friendly client report — HistoryRows component, the
  conditional score/drift/tells/history blocks, the @media print stylesheet
  preserved (raw CSS), onclick print button, and the literal '{ system: true }'
  kept as a string so JSX doesn't parse the braces.

Verified: landing.test.js (14/14, covers both pages' brand marks + report's
@media print), tsc clean.
